OPENEUROPA-3276: Adjust Kernel tests and add new test trait.
