+++ tests/Kernel/ListsSourceTest.php
@@ -5,16 +5,17 @@ declare(strict_types = 1);
 namespace Drupal\Tests\oe_list_pages\Kernel;
 
 use Drupal\Core\Plugin\PluginBase;
-use Drupal\facets\Entity\Facet;
-use Drupal\facets\FacetInterface;
 use Drupal\KernelTests\Core\Entity\EntityKernelTestBaseTest;
 use Drupal\oe_list_pages\ListSourceFactory;
 use Drupal\search_api\Entity\Index;
+use Drupal\Tests\oe_list_pages\Traits\SearchApiTestTrait;
 
 /**
  * Tests the List sources and their properties.
  */
 class ListsSourceTest extends EntityKernelTestBaseTest {
+
+  use SearchApiTestTrait;
 
   /**
    * {@inheritdoc}
@@ -159,32 +160,6 @@ class ListsSourceTest extends EntityKernelTestBaseTest {
     $this->assertArrayHasKey('category', $filters_item);
     $this->assertArrayNotHasKey('keywords', $filters_item);
     $this->assertArrayHasKey('width', $filters_item);
-  }
-
-  /**
-   * Creates a facet for the specified field.
-   *
-   * @param string $field
-   *   The field.
-   * @param string $search_id
-   *   The search id.
-   *
-   * @return \Drupal\facets\FacetInterface
-   *   The created facet.
-   */
-  private function createFacet(string $field, string $search_id): FacetInterface {
-    $entity = Facet::create([
-      // Id just needs to be unique when generated.
-      'id' => md5($search_id) . '_' . $field,
-      'name' => 'Facet for ' . $field,
-    ]);
-    $entity->setWidget('links');
-    $entity->setFieldIdentifier($field);
-    $entity->setEmptyBehavior(['behavior' => 'none']);
-    $entity->setFacetSourceId($search_id);
-    $entity->save();
-
-    return $entity;
   }
 
 }
